+++ cms/trunk/admin/changegroupperm.php
@@ -13,27 +13,38 @@ if (isset($_POST["cancel"])) {
 	return;
 }
 
-$db = new DB($config);
+$userid = get_userid();
+$access = check_permission($config, $userid, 'Modify Permissions');
 
-if (isset($_POST["changeperm"])) {
+if ($access) {
 
-	$query = "DELETE FROM ".$config->db_prefix."group_perms WHERE group_id = " . $group_id;
-	$result = $db->query($query);
+	$db = new DB($config);
 
-	foreach ($_POST as $key=>$value) {
-		if (strpos($key,"perm-") == 0) {
-			$query = "INSERT INTO ".$config->db_prefix."group_perms (group_id, permission_id, create_date, modified_date) VALUES (".mysql_real_escape_string($group_id).", ".mysql_real_escape_string(substr($key,5)).", now(), now())";
-			$result = $db->query($query);
+	if (isset($_POST["changeperm"])) {
+
+		$query = "DELETE FROM ".$config->db_prefix."group_perms WHERE group_id = " . $group_id;
+		$result = $db->query($query);
+
+		foreach ($_POST as $key=>$value) {
+			if (strpos($key,"perm-") == 0) {
+				$query = "INSERT INTO ".$config->db_prefix."group_perms (group_id, permission_id, create_date, modified_date) VALUES (".mysql_real_escape_string($group_id).", ".mysql_real_escape_string(substr($key,5)).", now(), now())";
+				$result = $db->query($query);
+			}
 		}
+
+		$db->close();
+		redirect("listgroups.php");
+		return;
+
 	}
-
-	$db->close();
-	redirect("listgroups.php");
-	return;
-
 }
 
 include_once("header.php");
+
+if (!$access) {
+	print "<h3>No Access to Modify Group Permissions</h3>";
+}
+else {
 
 ?>
 <h3>Permissions for group: Name</h3>
@@ -90,6 +101,8 @@ include_once("header.php");
 </form>
 
 <?php
+
+}
 
 include_once("footer.php");
