+++ cms/trunk/admin/deletesection.php
@@ -8,12 +8,16 @@ $section_id = -1;
 if (isset($_GET["section_id"])) {
 
 	$section_id = $_GET["section_id"];
+	$userid = get_userid();
+	$access = check_permission($config, $userid, 'Remove Section');
 
-	$db = new DB($config);
+	if ($access) {
+		$db = new DB($config);
 
-	$query = "DELETE FROM ".$config->db_prefix."sections where section_id = $section_id";
-	$result = $db->query($query);
-	$db->close();
+		$query = "DELETE FROM ".$config->db_prefix."sections where section_id = $section_id";
+		$result = $db->query($query);
+		$db->close();
+	}
 }
 
 redirect("listsections.php");
